월요일까지 과제 커밋

+++ src/main/java/com/example/hanghaememo/service/MemoService.java
@@ -7,6 +7,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
+import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
@@ -22,18 +23,34 @@ public class MemoService {
     public List<Memo> getMemos() {
         return memoRepository.findAllByOrderByModifiedAtDesc();
     }
+
     @Transactional
     public Long update(Long id, MemoRequestDto requestDto) {
+        long a = 100L;
+        Memo memos = new Memo(requestDto);
         Memo memo = memoRepository.findById(id).orElseThrow(
                 ()-> new IllegalArgumentException("아이디가 존재하지 않습니다")
         );
-        memo.update(requestDto);
+        if(Objects.equals(memo.getPassword(), requestDto.getPassword())){
+            memo.update(requestDto);
+        }else{
+            return a;
+        }
         return memo.getId();
     }
     @Transactional
-    public Long deleteMemo(Long id){
-        memoRepository.deleteById(id);
-        return id;
+    public Long deleteMemo(Long id,MemoRequestDto requestDto){
+        long a = 100L;
+
+        Memo memo = memoRepository.findById(id).orElseThrow(
+                ()-> new IllegalArgumentException("아이디가 존재하지 않습니다")
+        );
+        if (Objects.equals(memo.getPassword(), requestDto.getPassword())){
+            memoRepository.deleteById(id);
+            return memo.getId();
+        }else {
+            return a;
+        }
     }
 
 }
